+++ pages/moderator/AnalyticOrder.js
@@ -1,9 +1,10 @@
 import { Component } from "react";
-import { ScrollView, StyleSheet, Text, View } from "react-native";
+import { Pressable, ScrollView, StyleSheet, Text, View } from "react-native";
 import { BarChart } from 'react-native-charts-wrapper';
 import auth from '@react-native-firebase/auth';
 import database from '@react-native-firebase/database';
 import moment from "moment";
+
 
 class AnalyticOrder extends Component {
 
@@ -11,6 +12,8 @@ class AnalyticOrder extends Component {
 
         totalRevenue: 0,
         archivedOrdersaArr: [],
+        prodArr: [],
+        bestSeller: 0,
 
     }
 
@@ -25,6 +28,11 @@ class AnalyticOrder extends Component {
         this.setState({ archivedOrdersaArr })
 
     }
+
+    setProdArr = (arr) => {
+        this.setState({ prodArr: arr });
+    }
+
 
     getData = () => {
 
@@ -47,11 +55,48 @@ class AnalyticOrder extends Component {
                 }
             });
 
+        database()
+            .ref('/Products')
+            .on('value', (snapshot) => {
+                if (snapshot.exists()) {
+
+                    let data = [];
+                    snapshot.forEach((child) => {
+                        temp = child.val()
+                        temp.id = child.key
+                        data.push(temp)
+                        this.setProdArr(data)
+                    })
+                }
+                else {
+                    return null
+                }
+            });
+
 
     }
 
     componentDidMount() {
         this.getData()
+    }
+
+    getProdData(inp) {
+
+        data = this.state.prodArr
+        let temp = {}
+
+        for (i = 0; i < data.length; i++) {
+
+            if (data[i].id == inp) {
+
+                temp.name = data[i].product_name
+                temp.price = data[i].product_price
+                temp.image = data[i].product_image
+
+            }
+
+        }
+        return temp
     }
 
 
@@ -74,8 +119,6 @@ class AnalyticOrder extends Component {
                 sum = sum + archivedOrdersaArr[i].totalPayment
 
             }
-
-
         }
 
         return sum
@@ -87,11 +130,20 @@ class AnalyticOrder extends Component {
         let cartArr = []
         let sum = 0
         const { archivedOrdersaArr } = this.state
+        const currentDate = new Date();
+        const currentMonth = currentDate.getMonth()
+        const currentYear = currentDate.getFullYear()
 
         for (i = 0; i < archivedOrdersaArr.length; i++) {
 
-            cartArr.push(archivedOrdersaArr[i].cart)
+            let itemMonth = moment(archivedOrdersaArr[i].timestamp).month()
+            let itemYear = moment(archivedOrdersaArr[i].timestamp).year()
 
+            if (itemMonth === currentMonth && itemYear === currentYear) {
+
+                cartArr.push(archivedOrdersaArr[i].cart)
+
+            }
         }
 
         for (i = 0; i < cartArr.length; i++) {
@@ -105,6 +157,93 @@ class AnalyticOrder extends Component {
         }
 
         return sum
+
+    }
+
+    getBestSeller = () => {
+
+        const mergedArray = [];
+        let cartArr = []
+        let productArr = []
+        let result = { 'quantity': 0 }
+        const { archivedOrdersaArr } = this.state
+        const currentDate = new Date();
+        const currentMonth = currentDate.getMonth()
+        const currentYear = currentDate.getFullYear()
+
+        for (i = 0; i < archivedOrdersaArr.length; i++) {
+
+            let itemMonth = moment(archivedOrdersaArr[i].timestamp).month()
+            let itemYear = moment(archivedOrdersaArr[i].timestamp).year()
+
+            if (itemMonth === currentMonth && itemYear === currentYear) {
+
+                cartArr.push(archivedOrdersaArr[i].cart)
+
+            }
+        }
+        for (i = 0; i < cartArr.length; i++) {
+
+            for (j = 0; j < cartArr[i].length; j++) {
+
+                let product = {}
+                product.id = cartArr[i][j].item_id
+                product.quantity = cartArr[i][j].quantity
+                productArr.push(product)
+
+            }
+        }
+        productArr.forEach((obj) => {
+            const existingObj = mergedArray.find((item) => item.id === obj.id);
+
+            if (existingObj) {
+                existingObj.quantity += obj.quantity;
+            } else {
+                mergedArray.push(obj);
+            }
+        });
+
+        let biggest = mergedArray[0]
+
+        for (i = 0; i < mergedArray.length; i++) {
+
+            if (biggest.quantity < mergedArray[i].quantity) {
+
+                biggest = mergedArray[i]
+                result = biggest
+
+
+            }
+
+        }
+
+        return result
+
+    }
+
+    getOrdersPlaced = () => {
+
+        const { archivedOrdersaArr } = this.state
+        const currentDate = new Date();
+        const currentMonth = currentDate.getMonth()
+        const currentYear = currentDate.getFullYear()
+        let cartArr = []
+
+        for (i = 0; i < archivedOrdersaArr.length; i++) {
+
+            let itemMonth = moment(archivedOrdersaArr[i].timestamp).month()
+            let itemYear = moment(archivedOrdersaArr[i].timestamp).year()
+
+            if (itemMonth === currentMonth && itemYear === currentYear) {
+
+                cartArr.push(archivedOrdersaArr[i].cart)
+
+            }
+        }
+
+        let totalOrders = cartArr.length
+
+        return totalOrders
 
     }
 
@@ -126,14 +265,18 @@ class AnalyticOrder extends Component {
                             <Text>TOTAL REVENUE</Text>
                         </View>
                     </View>
+
                     <View style={styles.upperLayer}>
+
                         <View style={styles.upperSingleBox}>
-                            <Text style={styles.resultText}>RM 2800.50</Text>
-                            <Text>NET PROFIT</Text>
+                            <Text style={[styles.resultText, { fontSize: 20 }]}>{this.getBestSeller().id ? this.getProdData(this.getBestSeller().id).name : 'N/A'}</Text>
+                            {/* <Text>{this.getBestSeller().quantity}</Text> */}
+                            <Text>{'BEST SELLER (' + this.getBestSeller().quantity + ')'}</Text>
                         </View>
+
                         <View style={styles.upperSingleBox}>
-                            <Text style={styles.resultText}>23%</Text>
-                            <Text>PROFIT MARGIN</Text>
+                            <Text style={styles.resultText}>{this.getOrdersPlaced()}</Text>
+                            <Text>ORDERS PLACED</Text>
                         </View>
                     </View>
                 </View>
@@ -183,6 +326,7 @@ const styles = StyleSheet.create({
         fontWeight: 'bold',
         color: '#DB9B06',
         marginBottom: 10,
+        textAlign: 'center',
     },
     container: {
         flex: 1,